pyproject.toml
[tool.poetry]
name = "discord-monkey-bot"
version = "0.0.1"
description = "discord bot to display monkeys"
authors = ["Your Name <you@example.com>"]
readme = "README.rst"

packages = [
    { include = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.7.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
bot = "src.service:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

